Correct the ordering of after hooks for Jest 20.
The logic for running the after hooks included a work around for a bug that
existed in Jest 18 and 19.  The bug has since been fixed in Jest 20, so we're
removing the fix.  Because removing the workaround will break 18 and 19, this
is a breaking change, and as such I've dropped 18/19 references from
package.json and updated to 20 everywhere.