Add a semver switch to protect afterAll order in older versions of Jest.
There is a bug in Jest 18/19 that processes the afterAll hooks in the wrong
order. This bit of logic is meant to understand which method Jest is using
for moving through the list of afterAll hooks, and supply them in the order
that gets them applied in the order we want.